tools: move utils to awsrules package
