+++ go.mod
@@ -17,11 +17,11 @@ require (
 	github.com/jessevdk/go-flags v1.4.0
 	github.com/mattn/go-colorable v0.1.6
 	github.com/mitchellh/go-homedir v1.1.0
+	github.com/serenize/snaker v0.0.0-20171204205717-a683aaf2d516
 	github.com/sourcegraph/go-lsp v0.0.0-20181119182933-0c7d621186c1
 	github.com/sourcegraph/jsonrpc2 v0.0.0-20190106185902-35a74f039c6a
 	github.com/spf13/afero v1.3.0
 	github.com/terraform-linters/tflint-plugin-sdk v0.1.2-0.20200621094814-c5fcef3ab795
-	github.com/terraform-linters/tflint/tools v0.0.0-20200622212259-1b57b2fe8f33
 	github.com/terraform-providers/terraform-provider-aws v2.65.0+incompatible
 	github.com/zclconf/go-cty v1.5.0
 )
