+++ rules/awsrules/generator-utils/main.go
