+++ rules/awsrules/models/generator/rule.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/schema"
-	"github.com/terraform-linters/tflint/tools/utils"
+	utils "github.com/terraform-linters/tflint/rules/awsrules/generator-utils"
 )
 
 type ruleMeta struct {
